Moved throwPostgressException and fixed some includes
